Disable parser cache less often
USERLANGUAGECODE shouldn't disable the parser cache, since userlang
variation exists in the parser cache key when it is required.  To
achieve this, use $parser->getOptions()->getUserLang() instead of
$wgUser->getOption( 'language' ).

Additionally, there is no need to disable the parser cache for LOGO,
since wgLogo is a config variable, and its value is not dynamic.

Thanks to RheingoldRiver for this patch.

Change-Id: I853abf4749d3cf3028b750f89d28209188c23440