Add 7 new variables
{{PAGEIMAGE}} will output the main image of the current page
as determined by Extension:PageImages. If the extension is not
installed or no main image was determined then {{PAGEIMAGE}}
will output nothing.

{{REALNAME}}, when used in a user page, will output the real name
of the user. However, if used like this {{REALNAME:Jimbo}} it will
output the real name of the user Jimbo. This dual behavior is
similar to core variables like {{PAGENAME}}.

{{SUBPAGES}} will output the subpages of the current page,
in plain text, comma-separated and in alphabetic order, like so:
Root/Sub, Root/Sub2, Root/Sub2/SubSub

{{WHATLINKSHERE}} will output the titles of the pages that link
to the current page, comma-separated and in alphabetic order,
like so: Bar, Baz, Foo

{{FIRSTREVISIONID}} will output the ID of the first revision
of the current page. Identical to the core variable {{REVISIONID}}
but for the first revision instead of the current one.

{{FIRSTREVISIONTIMESTAMP}} will output the timestamp of the first
revision of the current page. Identical to the core variable
{{REVISIONTIMESTAMP}} but for the first revision instead of the
current one.

{{FIRSTREVISIONUSER}} will output the user name (without the User
prefix) or IP address associated to the first revision of the
current page. Identical to the core variable {{REVISIONUSER}}
but for the first revision instead of the current one.

Change-Id: Id0687cdfb0e7ad23d4caf3ca58a2af85b4f63a2d