Change method of disabling parser cache for {{CURRENTLOGGEDUSER}}
The goal is to disable parser cache but not CDN cache. The previous
solution used an article view hook. This was unideal because it
would often cause MediaWiki to parse the page twice, doubling the
load time, since it had to parse the page to see if the flag to
disable cache was present, and then parse the page later to show
it to the user. This would always happen during a cache reject, but
it seemed to also happen during ordinary cache misses sometimes, for
unclear reasons.

Instead we split the cache between logged in and logged out, but
always mark pages as uncacheable if the user is logged in. This
means that logged out users can still enjoy parser cache, and CDN
cache is still enabled.

Some other minor benefits are:
* The cache report HTML comment will now be correct, reducing confusion
* This will no longer save things in parser cache that cannot be used,
potentially leaving more room for important stuff.
* If you access a page not via the main view (e.g. via api) where the
old hook doesn't run, you could have gotten an incorrect page view.

[WIK-791]

Change-Id: I7b3d76df3d36a38adf9e4160f69d063d7167dc8e