Don't unnessarily disable cache
The Logo is not going to change, so no need to disable cache.

first revision can only be wrong if a preview or edit stash is being
used before the first revision is saved. After that point it is
safe to cache because you cannot go back and add a new first
revision.

Change-Id: I95689fd56dd305411c200084b4332e2ebb6d3602